requirements.txt
crewai>=0.11.0
pandas>=2.2.0
pyarrow>=15.0.0
duckdb>=0.9.2
langchain>=0.1.4
langchain-openai>=0.0.5
chromadb>=0.4.22
python-dotenv>=1.0.1
faker>=24.0.0
pydantic>=2.6.0
pytest>=8.0.0
pytest-asyncio>=0.23.5
langsmith>=0.1.5
loguru>=0.7.2
fastapi>=0.110.0
uvicorn>=0.27.0
python-multipart>=0.0.9
